# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bu, BeeHub Ders Seçim Botu için API dokümantasyonudur.",
        "title": "BeeHub Ders Seçim Botu API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Hello World",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/profile": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Hello World",
                "responses": {}
            }
        },
        "/beePicker/courses": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BeePicker"
                ],
                "summary": "Retrieves courses from the BeePicker.",
                "responses": {}
            }
        },
        "/beePicker/pick": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BeePicker"
                ],
                "summary": "Picks a course from the kepler.",
                "parameters": [
                    {
                        "description": "Request body containing the course codes",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/beepicker.pickRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Picking successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/beePicker/schedule": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BeePicker"
                ],
                "summary": "Retrieves schedules from the BeePicker.",
                "responses": {}
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BeePicker"
                ],
                "summary": "Selects a course from the BeePicker.",
                "parameters": [
                    {
                        "description": "Request body containing the ECRN",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/beepicker.scheduleSaveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Selection successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "beepicker.pickRequest": {
            "type": "object",
            "properties": {
                "courseCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "beepicker.scheduleSaveRequest": {
            "type": "object",
            "properties": {
                "ECRN": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "SCRN": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "scheduleName": {
                    "type": "string"
                }
            }
        }
    }
}